ml_service/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
spacy==3.7.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
language-tool-python==2.7.1
vosk==0.3.45
librosa==0.10.1
groq
numpy==1.26.0
pydub==0.25.1
pdfminer.six==20221105
python-dotenv
